% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fine_tuning_job.R
\name{create_fine_tuning_job}
\alias{create_fine_tuning_job}
\title{Create Fine-Tuning Job}
\usage{
create_fine_tuning_job(file_id, api_key)
}
\arguments{
\item{file_id}{The ID of the training file.}

\item{api_key}{The API key for authentication.}
}
\value{
The ID of the created fine-tuning job.
}
\description{
Create Fine-Tuning Job
}
