% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_request.R
\name{create_request}
\alias{create_request}
\title{Title}
\usage{
create_request(url, api_key, body = NULL, method = "GET")
}
\arguments{
\item{url}{string. The URL for the request.}

\item{api_key}{string. The API key for authentication.}

\item{body}{string. The body of the request.}

\item{method}{string. The method of the request, GET,DELETES}
}
\description{
Title
}
