% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tune_model.R
\name{fine_tune_model}
\alias{fine_tune_model}
\title{Title}
\usage{
fine_tune_model(file_id, api_key, interval = 10)
}
\arguments{
\item{file_id}{The ID of the training file.}

\item{api_key}{The API key for authentication.}

\item{interval}{The interval (in seconds) at which to check the fine-tuning job status.}
}
\description{
Title
}
