% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fine_tuned_models.R
\name{list_fine_tuned_models}
\alias{list_fine_tuned_models}
\title{Title}
\usage{
list_fine_tuned_models(api_key)
}
\arguments{
\item{api_key}{The API key for authentication.}
}
\description{
Title
}
