% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_delete_request.R
\name{perform_delete_request}
\alias{perform_delete_request}
\title{Perform Delete Request}
\usage{
perform_delete_request(url, api_key, model_id)
}
\arguments{
\item{url}{The URL for the DELETE request.}

\item{api_key}{The API key for authentication.}

\item{model_id}{The ID of the model to be deleted.}
}
\description{
Perform Delete Request
}
