% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{query_model}
\alias{query_model}
\title{Title}
\usage{
query_model(
  prompt,
  api_key,
  model_name = "gpt-4",
  max_tokens = NULL,
  temperature = 0,
  debug = FALSE
)
}
\arguments{
\item{prompt}{string. The prompt to send to the model}

\item{api_key}{string. The API key for authentication}

\item{model_name}{string. The name of the model to query}

\item{max_tokens}{integer. he maximum number of tokens to generate}

\item{temperature}{integer. The temperature for sampling}

\item{debug}{boolean. Whether to print debug information}
}
\description{
Title
}
