% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_training_data.R
\name{upload_training_data}
\alias{upload_training_data}
\title{Title}
\usage{
upload_training_data(data, api_key)
}
\arguments{
\item{data}{data.table. The data.table containing the user and assistant data.}

\item{api_key}{string. The API key for authentication.}
}
\description{
Title
}
