% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fine_tuning_job.R
\name{create_fine_tuning_job}
\alias{create_fine_tuning_job}
\title{Create Fine-Tuning Job}
\usage{
create_fine_tuning_job(file_id, api_key, debug = FALSE)
}
\arguments{
\item{file_id}{string. The ID of the training file.}

\item{api_key}{string. The API key for authentication.}

\item{debug}{boolean. Logical indicating whether to print debug information.}
}
\value{
The ID of the created fine-tuning job.
}
\description{
Create Fine-Tuning Job
}
