% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_fine_tune_model.R
\name{delete_fine_tune_model}
\alias{delete_fine_tune_model}
\title{Delete Fine-Tuned Models}
\usage{
delete_fine_tune_model(model_ids = NULL, api_key)
}
\arguments{
\item{model_ids}{A vector of model IDs to delete. If NULL, deletes all models except the latest.}

\item{api_key}{The API key for authentication.}
}
\description{
Delete Fine-Tuned Models
}
