% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{query_model}
\alias{query_model}
\title{Title}
\usage{
query_model(
  prompt,
  api_key,
  model_name = "gpt-4o",
  max_tokens = NULL,
  temperature = 0
)
}
\arguments{
\item{prompt}{The prompt to send to the model}

\item{api_key}{The API key for authentication}

\item{model_name}{The name of the model to query}

\item{max_tokens}{The maximum number of tokens to generate}

\item{temperature}{The temperature for sampling}
}
\description{
Title
}
