% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_training_data.R
\name{upload_training_data}
\alias{upload_training_data}
\title{Title}
\usage{
upload_training_data(df, api_key)
}
\arguments{
\item{df}{data.table}

\item{api_key}{The API key for authentication.}
}
\description{
Title
}
